+++ project_electron/aurora/urls.py
@@ -24,8 +24,7 @@ urlpatterns = [
     url(r'^app/$',              MainView.as_view(), name='app_home'),
     url(r'^app/transfers/',     include('orgs.transfers.urls')),
     url(r'^app/orgs/',          include('orgs.urls')),
-    url(r'^app/users/',         include('orgs.user_urls')),
-    url(r'^app/password/',      include('orgs.users.urls')),
+    url(r'^app/users/',         include('orgs.users.urls')),
     url(r'^app/accession/',     include('orgs.accession.urls')),
     url(r'^app/appraise/',      include('orgs.appraise.urls')),
     url(r'^app/rights/',        include('orgs.rights.urls')),
