--- project_electron/orgs/formatmixins.py
@@ -1,21 +1,0 @@
-import csv
-import datetime
-
-from django.http import HttpResponse
-
-class CSVResponseMixin(object):
-    csv_filename = 'transfers-{}.csv'.format(datetime.datetime.now())
-
-    def get_csv_filename(self):
-        return self.csv_filename
-
-    def render_to_csv(self, data):
-        response = HttpResponse(content_type='text/csv')
-        cd = 'attachment; filename="{0}"'.format(self.get_csv_filename())
-        response['Content-Disposition'] = cd
-
-        writer = csv.writer(response)
-        for row in data:
-            writer.writerow(row)
-
-        return response
