+++ project_electron/orgs/mixins/formatmixins.py
@@ -1,0 +1,43 @@
+import csv
+import datetime
+
+from django.http import HttpResponse, JsonResponse
+
+
+class CSVResponseMixin(object):
+    csv_filename = 'transfers-{}.csv'.format(datetime.datetime.now())
+
+    def get_csv_filename(self):
+        return self.csv_filename
+
+    def render_to_csv(self, data):
+        response = HttpResponse(content_type='text/csv')
+        cd = 'attachment; filename="{0}"'.format(self.get_csv_filename())
+        response['Content-Disposition'] = cd
+
+        writer = csv.writer(response)
+        for row in data:
+            writer.writerow(row)
+
+        return response
+
+
+class JSONResponseMixin:
+    """
+    A mixin that can be used to render a JSON response.
+    """
+    def render_to_json_response(self, context, **response_kwargs):
+        """
+        Returns a JSON response, transforming 'context' to make the payload.
+        """
+        return JsonResponse(
+            self.get_data(context),
+            **response_kwargs
+        )
+
+    def get_data(self, context):
+        """
+        Returns an object that will be serialized as JSON by json.dumps().
+        """
+
+        return context
