+++ project_electron/orgs/rights/views.py
@@ -8,11 +8,11 @@ from django.views.generic import CreateView, UpdateView, DetailView, TemplateVie
 from orgs.rights.models import *
 from orgs.rights.forms import *
 from orgs.models import BagItProfile, BagItProfileBagInfo, BagItProfileBagInfoValues
-from orgs.authmixins import *
-from orgs.mixins import JSONResponseMixin
+from orgs.mixins.authmixins import *
+from orgs.mixins.formatmixins import JSONResponseMixin
 
 from django.shortcuts import render, redirect, render_to_response, get_object_or_404
-from orgs.authmixins import OrgReadViewMixin
+from orgs.mixins.authmixins import OrgReadViewMixin
 
 class RightsManageView(ManagingArchivistMixin, CreateView):
     template_name = 'rights/manage.html'
