+++ project_electron/orgs/templates/orgs/update.html
@@ -62,5 +62,5 @@
 
 
 {% block extra_js %}
-<script type="text/javascript" src="{% static 'js/rights_js.js' %}"></script>
+<script type="text/javascript" src="{% static 'js/delete_object_js.js' %}"></script>
 {% endblock %}
