+++ project_electron/orgs/templates/transfers/transfer_detail.html
@@ -148,5 +148,5 @@
 
 
 {% block extra_js %}
-<script type="text/javascript" src="{% static 'js/rights_js.js' %}"></script>
+<script type="text/javascript" src="{% static 'js/delete_object_js.js' %}"></script>
 {% endblock %}
